requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.3
cycler==0.12.1
distlib==0.3.8
dj-database-url==3.0.1
Django==5.1.2
djangorestframework==3.16.1
dnspython==2.7.0
fastapi==0.116.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.59.2
fsspec==2025.9.0
gengo==1.0.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.9
manage.py==0.2.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
myapp==0.1.dev0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.1.0
platformdirs==4.3.6
polars==1.33.0
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyjokes==0.8.3
pymongo==4.11.2
pyparsing==3.2.3
pypiwin32==223
pyproject_hooks==1.2.0
pythin==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.1
starlette==0.47.2
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
ultralytics==8.3.196
ultralytics-thop==2.0.17
urllib3==2.2.3
uvicorn==0.35.0
virtualenv==20.26.6
Werkzeug==3.1.3
whitenoise==6.10.0
